Lagt inn KafkaProducer
