+++ src/main/kotlin/no/nav/helsearbeidsgiver/plugins/Routing.kt
@@ -4,16 +4,15 @@ import io.ktor.http.HttpStatusCode
 import io.ktor.server.application.Application
 import io.ktor.server.application.call
 import io.ktor.server.auth.authenticate
-import io.ktor.server.auth.principal
 import io.ktor.server.plugins.swagger.swaggerUI
 import io.ktor.server.response.respond
 import io.ktor.server.response.respondText
 import io.ktor.server.routing.get
 import io.ktor.server.routing.routing
+import no.nav.helsearbeidsgiver.auth.getConsumerOrgnr
+import no.nav.helsearbeidsgiver.auth.tokenValidationContext
 import no.nav.helsearbeidsgiver.forespoersel.ForespoerselService
-import no.nav.helsearbeidsgiver.getConsumerOrgnr
 import no.nav.helsearbeidsgiver.inntektsmelding.InntektsmeldingService
-import no.nav.security.token.support.v2.TokenValidationContextPrincipal
 import org.slf4j.LoggerFactory
 
 private val forespoerselService = ForespoerselService()
@@ -31,17 +30,14 @@ fun Application.configureRouting() {
                 call.respond(forespoerselService.hentForespoersler())
             }
             get("/inntektsmeldinger") {
-                val principal = call.principal<TokenValidationContextPrincipal>()
-                val tokenValidationContext = principal?.context
-                if (tokenValidationContext != null) {
-                    val consumerOrgnr = tokenValidationContext.getConsumerOrgnr()
-                    consumerOrgnr?.let {
-                        LOG.info("Consumer orgnr: $it")
+                val consumerOrgnr = tokenValidationContext().getConsumerOrgnr()
 
-                        call.respond(inntektsmeldingService.hentInntektsmeldingerByOrgNr(consumerOrgnr))
-                    }
+                if (consumerOrgnr != null) {
+                    LOG.info("Henter inntektsmeldinger for orgnr: $consumerOrgnr")
+                    call.respond(inntektsmeldingService.hentInntektsmeldingerByOrgNr(consumerOrgnr))
                 } else {
-                    call.respondText("TokenValidationContext not found", status = HttpStatusCode.Unauthorized)
+                    LOG.warn("Consumer orgnr mangler")
+                    call.respond(HttpStatusCode.BadRequest, "Consumer orgnr mangler")
                 }
             }
         }
